Move query and tx commands to modules
