+++ cmd/gaia/cmd/gaiacli/query.go
@@ -21,51 +21,6 @@ func queryCmd(cdc *amino.Codec) *cobra.Command {
 		Short:   "Querying subcommands",
 	}
 
-	// Group staking queries under a subcommand
-	stakeQueryCmd := &cobra.Command{
-		Use:   "stake",
-		Short: "Querying commands for the staking module",
-	}
-
-	stakeQueryCmd.AddCommand(client.GetCommands(
-		stakecmd.GetCmdQueryDelegation(storeStake, cdc),
-		stakecmd.GetCmdQueryDelegations(storeStake, cdc),
-		stakecmd.GetCmdQueryUnbondingDelegation(storeStake, cdc),
-		stakecmd.GetCmdQueryUnbondingDelegations(storeStake, cdc),
-		stakecmd.GetCmdQueryRedelegation(storeStake, cdc),
-		stakecmd.GetCmdQueryRedelegations(storeStake, cdc),
-		stakecmd.GetCmdQueryValidator(storeStake, cdc),
-		stakecmd.GetCmdQueryValidators(storeStake, cdc),
-		stakecmd.GetCmdQueryValidatorDelegations(storeStake, cdc),
-		stakecmd.GetCmdQueryValidatorUnbondingDelegations(queryRouteStake, cdc),
-		stakecmd.GetCmdQueryValidatorRedelegations(queryRouteStake, cdc),
-		stakecmd.GetCmdQueryParams(storeStake, cdc),
-		stakecmd.GetCmdQueryPool(storeStake, cdc))...)
-
-	// Group gov queries under a subcommand
-	govQueryCmd := &cobra.Command{
-		Use:   "gov",
-		Short: "Querying commands for the governance module",
-	}
-
-	govQueryCmd.AddCommand(client.GetCommands(
-		govcmd.GetCmdQueryProposal(storeGov, cdc),
-		govcmd.GetCmdQueryProposals(storeGov, cdc),
-		govcmd.GetCmdQueryVote(storeGov, cdc),
-		govcmd.GetCmdQueryVotes(storeGov, cdc),
-		govcmd.GetCmdQueryParams(storeGov, cdc),
-		govcmd.GetCmdQueryDeposit(storeGov, cdc),
-		govcmd.GetCmdQueryDeposits(storeGov, cdc))...)
-
-	// Group slashing queries under a subcommand
-	slashingQueryCmd := &cobra.Command{
-		Use:   "slashing",
-		Short: "Querying commands for the slashing module",
-	}
-
-	slashingQueryCmd.AddCommand(client.GetCommands(
-		slashingcmd.GetCmdQuerySigningInfo(storeSlashing, cdc))...)
-
 	// Query commcmmand structure
 	queryCmd.AddCommand(
 		rpc.BlockCommand(),
@@ -74,9 +29,9 @@ func queryCmd(cdc *amino.Codec) *cobra.Command {
 		tx.QueryTxCmd(cdc),
 		client.LineBreak,
 		client.GetCommands(authcmd.GetAccountCmd(storeAcc, cdc, authcmd.GetAccountDecoder(cdc)))[0],
-		stakeQueryCmd,
-		govQueryCmd,
-		slashingQueryCmd,
+		stakecmd.GetQueryCmd(storeStake, cdc),
+		govcmd.GetQueryCmd(storeGov, cdc),
+		slashingcmd.GetQueryCmd(storeSlashing, cdc),
 	)
 
 	return queryCmd
