+++ x/gov/client/cli/tx.go
@@ -15,7 +15,7 @@ import (
 	"io/ioutil"
 	"strings"
 
-	"github.com/cosmos/cosmos-sdk/x/gov/client"
+	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -46,6 +46,22 @@ var proposalFlags = []string{
 	flagDescription,
 	flagProposalType,
 	flagDeposit,
+}
+
+// GetTxCmd returns the transaction commands for this module
+func GetTxCmd(storeKey string, cdc *codec.Codec) *cobra.Command {
+	govTxCmd := &cobra.Command{
+		Use:   "gov",
+		Short: "Governance transactions subcommands",
+	}
+
+	govTxCmd.AddCommand(client.PostCommands(
+		GetCmdDeposit(cdc),
+		GetCmdVote(cdc),
+		GetCmdSubmitProposal(cdc),
+	)...)
+
+	return govTxCmd
 }
 
 // GetCmdSubmitProposal implements submitting a proposal transaction command.
@@ -130,7 +146,7 @@ func parseSubmitProposalFlags() (*proposal, error) {
 	if proposalFile == "" {
 		proposal.Title = viper.GetString(flagTitle)
 		proposal.Description = viper.GetString(flagDescription)
-		proposal.Type = client.NormalizeProposalType(viper.GetString(flagProposalType))
+		proposal.Type = govgovClient.NormalizeProposalType(viper.GetString(flagProposalType))
 		proposal.Deposit = viper.GetString(flagDeposit)
 		return proposal, nil
 	}
@@ -218,7 +234,7 @@ func GetCmdVote(cdc *codec.Codec) *cobra.Command {
 			proposalID := uint64(viper.GetInt64(flagProposalID))
 			option := viper.GetString(flagOption)
 
-			byteVoteOption, err := gov.VoteOptionFromString(client.NormalizeVoteOption(option))
+			byteVoteOption, err := gov.VoteOptionFromString(govgovClient.NormalizeVoteOption(option))
 			if err != nil {
 				return err
 			}
@@ -245,317 +261,6 @@ func GetCmdVote(cdc *codec.Codec) *cobra.Command {
 
 	cmd.Flags().String(flagProposalID, "", "proposalID of proposal voting on")
 	cmd.Flags().String(flagOption, "", "vote option {yes, no, no_with_veto, abstain}")
-
-	return cmd
-}
-
-// GetCmdQueryProposal implements the query proposal command.
-func GetCmdQueryParams(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "param [param-type]",
-		Short: "Query the parameters (voting|tallying|deposit) of the governance process",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			paramType := args[0]
-
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/params/%s", queryRoute, paramType), nil)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(string(res))
-			return nil
-		},
-	}
-
-	return cmd
-}
-
-// GetCmdQueryProposal implements the query proposal command.
-func GetCmdQueryProposal(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "proposal",
-		Short: "Query details of a single proposal",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			proposalID := uint64(viper.GetInt64(flagProposalID))
-
-			params := gov.QueryProposalParams{
-				ProposalID: proposalID,
-			}
-
-			bz, err := cdc.MarshalJSON(params)
-			if err != nil {
-				return err
-			}
-
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/proposal", queryRoute), bz)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(string(res))
-			return nil
-		},
-	}
-
-	cmd.Flags().String(flagProposalID, "", "proposalID of proposal being queried")
-
-	return cmd
-}
-
-// GetCmdQueryProposals implements a query proposals command.
-func GetCmdQueryProposals(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "proposals",
-		Short: "Query proposals with optional filters",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			bechDepositerAddr := viper.GetString(flagDepositer)
-			bechVoterAddr := viper.GetString(flagVoter)
-			strProposalStatus := viper.GetString(flagStatus)
-			numLimit := uint64(viper.GetInt64(flagNumLimit))
-
-			params := gov.QueryProposalsParams{
-				Limit: numLimit,
-			}
-
-			if len(bechDepositerAddr) != 0 {
-				depositerAddr, err := sdk.AccAddressFromBech32(bechDepositerAddr)
-				if err != nil {
-					return err
-				}
-				params.Depositer = depositerAddr
-			}
-
-			if len(bechVoterAddr) != 0 {
-				voterAddr, err := sdk.AccAddressFromBech32(bechVoterAddr)
-				if err != nil {
-					return err
-				}
-				params.Voter = voterAddr
-			}
-
-			if len(strProposalStatus) != 0 {
-				proposalStatus, err := gov.ProposalStatusFromString(client.NormalizeProposalStatus(strProposalStatus))
-				if err != nil {
-					return err
-				}
-				params.ProposalStatus = proposalStatus
-			}
-
-			bz, err := cdc.MarshalJSON(params)
-			if err != nil {
-				return err
-			}
-
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/proposals", queryRoute), bz)
-			if err != nil {
-				return err
-			}
-
-			var matchingProposals []gov.Proposal
-			err = cdc.UnmarshalJSON(res, &matchingProposals)
-			if err != nil {
-				return err
-			}
-
-			if len(matchingProposals) == 0 {
-				fmt.Println("No matching proposals found")
-				return nil
-			}
-
-			for _, proposal := range matchingProposals {
-				fmt.Printf("  %d - %s\n", proposal.GetProposalID(), proposal.GetTitle())
-			}
-
-			return nil
-		},
-	}
-
-	cmd.Flags().String(flagNumLimit, "", "(optional) limit to latest [number] proposals. Defaults to all proposals")
-	cmd.Flags().String(flagDepositer, "", "(optional) filter by proposals deposited on by depositer")
-	cmd.Flags().String(flagVoter, "", "(optional) filter by proposals voted on by voted")
-	cmd.Flags().String(flagStatus, "", "(optional) filter proposals by proposal status, status: deposit_period/voting_period/passed/rejected")
-
-	return cmd
-}
-
-// Command to Get a Proposal Information
-// GetCmdQueryVote implements the query proposal vote command.
-func GetCmdQueryVote(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "vote",
-		Short: "Query details of a single vote",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			proposalID := uint64(viper.GetInt64(flagProposalID))
-
-			voterAddr, err := sdk.AccAddressFromBech32(viper.GetString(flagVoter))
-			if err != nil {
-				return err
-			}
-
-			params := gov.QueryVoteParams{
-				Voter:      voterAddr,
-				ProposalID: proposalID,
-			}
-			bz, err := cdc.MarshalJSON(params)
-			if err != nil {
-				return err
-			}
-
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/vote", queryRoute), bz)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(string(res))
-			return nil
-		},
-	}
-
-	cmd.Flags().String(flagProposalID, "", "proposalID of proposal voting on")
-	cmd.Flags().String(flagVoter, "", "bech32 voter address")
-
-	return cmd
-}
-
-// GetCmdQueryVotes implements the command to query for proposal votes.
-func GetCmdQueryVotes(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "votes",
-		Short: "Query votes on a proposal",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			proposalID := uint64(viper.GetInt64(flagProposalID))
-
-			params := gov.QueryVotesParams{
-				ProposalID: proposalID,
-			}
-			bz, err := cdc.MarshalJSON(params)
-			if err != nil {
-				return err
-			}
-
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/votes", queryRoute), bz)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(string(res))
-			return nil
-		},
-	}
-
-	cmd.Flags().String(flagProposalID, "", "proposalID of which proposal's votes are being queried")
-
-	return cmd
-}
-
-// Command to Get a specific Deposit Information
-// GetCmdQueryDeposit implements the query proposal deposit command.
-func GetCmdQueryDeposit(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "deposit",
-		Short: "Query details of a deposit",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			proposalID := uint64(viper.GetInt64(flagProposalID))
-
-			depositerAddr, err := sdk.AccAddressFromBech32(viper.GetString(flagDepositer))
-			if err != nil {
-				return err
-			}
-
-			params := gov.QueryDepositParams{
-				Depositer:  depositerAddr,
-				ProposalID: proposalID,
-			}
-			bz, err := cdc.MarshalJSON(params)
-			if err != nil {
-				return err
-			}
-
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/deposit", queryRoute), bz)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(string(res))
-			return nil
-		},
-	}
-
-	cmd.Flags().String(flagProposalID, "", "proposalID of proposal deposited on")
-	cmd.Flags().String(flagDepositer, "", "bech32 depositer address")
-
-	return cmd
-}
-
-// GetCmdQueryDeposits implements the command to query for proposal deposits.
-func GetCmdQueryDeposits(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "deposits",
-		Short: "Query deposits on a proposal",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			proposalID := uint64(viper.GetInt64(flagProposalID))
-
-			params := gov.QueryDepositsParams{
-				ProposalID: proposalID,
-			}
-			bz, err := cdc.MarshalJSON(params)
-			if err != nil {
-				return err
-			}
-
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/deposits", queryRoute), bz)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(string(res))
-			return nil
-		},
-	}
-
-	cmd.Flags().String(flagProposalID, "", "proposalID of which proposal's deposits are being queried")
-
-	return cmd
-}
-
-// GetCmdQueryTally implements the command to query for proposal tally result.
-func GetCmdQueryTally(queryRoute string, cdc *codec.Codec) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "tally",
-		Short: "Get the tally of a proposal vote",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliCtx := context.NewCLIContext().WithCodec(cdc)
-			proposalID := uint64(viper.GetInt64(flagProposalID))
-
-			params := gov.QueryTallyParams{
-				ProposalID: proposalID,
-			}
-			bz, err := cdc.MarshalJSON(params)
-			if err != nil {
-				return err
-			}
-
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/tally", queryRoute), bz)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println(string(res))
-			return nil
-		},
-	}
-
-	cmd.Flags().String(flagProposalID, "", "proposalID of which proposal is being tallied")
 
 	return cmd
 }
