+++ x/stake/client/cli/query.go
@@ -7,12 +7,37 @@ import (
 	"github.com/spf13/viper"
 	"github.com/tendermint/tendermint/libs/cli"
 
+	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/stake"
 	"github.com/cosmos/cosmos-sdk/x/stake/types"
 )
+
+// GetQueryCmd returns the query commands for this module
+func GetQueryCmd(storeKey string, cdc *codec.Codec) *cobra.Command {
+	stakeQueryCmd := &cobra.Command{
+		Use:   "stake",
+		Short: "Querying commands for the staking module",
+	}
+	stakeQueryCmd.AddCommand(client.GetCommands(
+		GetCmdQueryDelegation(storeKey, cdc),
+		GetCmdQueryDelegations(storeKey, cdc),
+		GetCmdQueryUnbondingDelegation(storeKey, cdc),
+		GetCmdQueryUnbondingDelegations(storeKey, cdc),
+		GetCmdQueryRedelegation(storeKey, cdc),
+		GetCmdQueryRedelegations(storeKey, cdc),
+		GetCmdQueryValidator(storeKey, cdc),
+		GetCmdQueryValidators(storeKey, cdc),
+		GetCmdQueryValidatorDelegations(storeKey, cdc),
+		GetCmdQueryValidatorUnbondingDelegations(storeKey, cdc),
+		GetCmdQueryValidatorRedelegations(storeKey, cdc),
+		GetCmdQueryParams(storeKey, cdc),
+		GetCmdQueryPool(storeKey, cdc))...)
+
+	return stakeQueryCmd
+}
 
 // GetCmdQueryValidator implements the validator query command.
 func GetCmdQueryValidator(storeName string, cdc *codec.Codec) *cobra.Command {
@@ -115,7 +140,7 @@ func GetCmdQueryValidators(storeName string, cdc *codec.Codec) *cobra.Command {
 }
 
 // GetCmdQueryValidatorUnbondingDelegations implements the query all unbonding delegatations from a validator command.
-func GetCmdQueryValidatorUnbondingDelegations(queryRoute string, cdc *codec.Codec) *cobra.Command {
+func GetCmdQueryValidatorUnbondingDelegations(storeKey string, cdc *codec.Codec) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "unbonding-delegations-from [operator-addr]",
 		Short: "Query all unbonding delegatations from a validator",
@@ -135,7 +160,7 @@ func GetCmdQueryValidatorUnbondingDelegations(queryRoute string, cdc *codec.Code
 			}
 
 			res, err := cliCtx.QueryWithData(
-				fmt.Sprintf("custom/%s/validatorUnbondingDelegations", queryRoute),
+				fmt.Sprintf("custom/%s/validatorUnbondingDelegations", storeKey),
 				bz)
 			if err != nil {
 				return err
@@ -150,7 +175,7 @@ func GetCmdQueryValidatorUnbondingDelegations(queryRoute string, cdc *codec.Code
 }
 
 // GetCmdQueryValidatorRedelegations implements the query all redelegatations from a validator command.
-func GetCmdQueryValidatorRedelegations(queryRoute string, cdc *codec.Codec) *cobra.Command {
+func GetCmdQueryValidatorRedelegations(storeKey string, cdc *codec.Codec) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "redelegations-from [operator-addr]",
 		Short: "Query all outgoing redelegatations from a validator",
@@ -170,7 +195,7 @@ func GetCmdQueryValidatorRedelegations(queryRoute string, cdc *codec.Codec) *cob
 			}
 
 			res, err := cliCtx.QueryWithData(
-				fmt.Sprintf("custom/%s/validatorRedelegations", queryRoute),
+				fmt.Sprintf("custom/%s/validatorRedelegations", storeKey),
 				bz)
 			if err != nil {
 				return err
@@ -288,7 +313,7 @@ func GetCmdQueryDelegations(storeName string, cdc *codec.Codec) *cobra.Command {
 
 // GetCmdQueryValidatorDelegations implements the command to query all the
 // delegations to a specific validator.
-func GetCmdQueryValidatorDelegations(queryRoute string, cdc *codec.Codec) *cobra.Command {
+func GetCmdQueryValidatorDelegations(storeKey string, cdc *codec.Codec) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "delegations-to [validator-addr]",
 		Short: "Query all delegations made to one validator",
@@ -308,7 +333,7 @@ func GetCmdQueryValidatorDelegations(queryRoute string, cdc *codec.Codec) *cobra
 
 			cliCtx := context.NewCLIContext().WithCodec(cdc)
 
-			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/validatorDelegations", queryRoute), bz)
+			res, err := cliCtx.QueryWithData(fmt.Sprintf("custom/%s/validatorDelegations", storeKey), bz)
 			if err != nil {
 				return err
 			}
